+++ src/containers/App/App.tsx
@@ -1,20 +1,18 @@
 import React, { useEffect } from 'react';
+import { ToastContainer } from 'react-toastify';
 import { CssBaseline } from '@material-ui/core';
 import { ThemeProvider, StylesProvider } from '@material-ui/styles';
+
 import { BreakpointsProvider } from 'hooks/useBreakpoints';
 import { darkTheme, lightTheme } from 'theme';
 import { Layout, Routes } from 'containers';
 import { useShallowSelector } from 'hooks';
 import userSelector from 'store/user/selectors';
-import { State, UserState } from 'types';
-import { ToastContainer } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
 import { useWalletConnectorContext } from 'services';
+import 'react-toastify/dist/ReactToastify.css';
 
 function App() {
-  const {
-    isLight, address, wallet,
-  } = useShallowSelector<State, UserState>(userSelector.getUser);
+  const { isLight, address, wallet } = useShallowSelector(userSelector.getUser);
   const { connect } = useWalletConnectorContext();
 
   const selectedTheme = isLight ? lightTheme : darkTheme;
